requirements.txt
pyworld==0.3.0
librosa==0.8.0
soundfile==0.10.3.post1
numba==0.48.0
scipy==1.10.1
resampy==0.2.2
tqdm==4.66.5
hydra-core==1.0.0
numpy==1.24.3
kaldiio==2.15.1